# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NazLibra POS API",
        "description": "Offline-first REST API for NazLibra POS mobile application. All protected routes require Bearer token + X-Tenant-Slug header.",
        "contact": {
            "email": "support@omnevo.net"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/v1/pos/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "List all held (parked) POS tickets",
                "description": "GET /api/v1/pos/tickets — list all held tickets for this tenant.",
                "operationId": "ticketIndex",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Location-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of held tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "tickets": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create or update a held POS ticket (parked cart)",
                "description": "POST /api/v1/pos/tickets — create or update a held ticket.\n\nThe client should upsert by sending the same `number` to update.\nIf `number` is omitted a new ticket is created.",
                "operationId": "ticketStore",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Location-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cart"
                                ],
                                "properties": {
                                    "number": {
                                        "description": "Existing ticket number to update",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "contact_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "cart": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "item_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "nullable": true
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "subtotal_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "discount_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "note": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created or updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "ticket": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pos/tickets/{ticket}": {
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete a held POS ticket",
                "description": "DELETE /api/v1/pos/tickets/{ticket} — discard a held ticket.",
                "operationId": "ticketDestroy",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Location-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}